+++ drivers/resdiskzpool/main.go
@@ -206,7 +206,7 @@ func (t *T) genHostID() error {
 
 // UnprovisionStop skips the normal pre-unprovision resource stop,
 // because zfs can only destroy imported pools. The Unprovision func
-// imports anyway, but if we don't export unecessary export/import is
+// imports anyway, but if we don't export unnecessary export/import is
 // saved.
 func (t *T) UnprovisionStop(ctx context.Context) error {
 	t.Log().Debugf("bypass export for unprovision")