+++ drivers/resfshost/main.go
@@ -385,7 +385,7 @@ func (t *T) ProvisionLeader(ctx context.Context) error {
 	if v, err := i1.IsFormated(devpath); err != nil {
 		t.Log().Warnf("skip mkfs: %s", err)
 	} else if v {
-		t.Log().Infof("%s is already formated", t.Device)
+		t.Log().Infof("%s is already formatted", t.Device)
 		return nil
 	}
 	i2, ok := fs.(filesystems.MKFSer)